+++ core/src/main/java/com/alibaba/fastjson2/writer/FieldWriter.java
@@ -707,7 +707,7 @@ public abstract class FieldWriter<T>
             return;
         }
 
-        final int SECONDS_PER_DAY = 60 * 60 * 24;
+        final long SECONDS_PER_DAY = 60 * 60 * 24;
 
         JSONWriter.Context ctx = jsonWriter.context;
         if (isDateFormatMillis() || ctx.isDateFormatMillis()) {
